add Camera's methods

+++ isceneobject.cpp
@@ -6,10 +6,10 @@ void ISceneObject::clean() {
     locatingPointsProjected.clear();
 }
 
-void ISceneObject::pushLocatingPointProjected(const cv::Vec2f &point) {
+void ISceneObject::pushLocatingPointProjected(const cv::Vec2d &point) {
     locatingPointsProjected.push_back(point);
 }
 
-const std::vector<cv::Vec3f>& ISceneObject::getLocatingPoints() const {
+const std::vector<cv::Vec3d>& ISceneObject::getLocatingPoints() const {
     return locatingPoints;
 }
